+++ src/Blameable/Mapping/Driver/Yaml.php
@@ -61,6 +61,10 @@ class Yaml extends File implements Driver
                         throw new InvalidMappingException("Field - [{$field}] trigger 'on' is not one of [update, create, change] in class - {$meta->getName()}");
                     }
 
+                    if (isset($mappingProperty['setterMethod'])) {
+                        $this->setSetterMethod($field, $mappingProperty['setterMethod'], $config);
+                    }
+
                     if ('change' === $mappingProperty['on']) {
                         if (!isset($mappingProperty['field'])) {
                             throw new InvalidMappingException("Missing parameters on property - {$field}, field must be set on [change] trigger in class - {$meta->getName()}");